+++ python/python/tests/test_dataset.py
@@ -3,7 +3,6 @@
 
 import base64
 import contextlib
-import copy
 import os
 import pickle
 import platform
@@ -32,15 +31,6 @@ from lance._dataset.sharded_batch_iterator import ShardedBatchIterator
 from lance.commit import CommitConflictError
 from lance.debug import format_fragment
 from lance.util import validate_vector_index
-
-CONFIG = {
-    "allow_http": "true",
-    "aws_access_key_id": "ACCESSKEY",
-    "aws_secret_access_key": "SECRETKEY",
-    "aws_endpoint": "http://localhost:9000",
-    "dynamodb_endpoint": "http://localhost:8000",
-    "aws_region": "us-west-2",
-}
 
 # Various valid inputs for write_dataset
 input_schema = pa.schema([pa.field("a", pa.float64()), pa.field("b", pa.int64())])
@@ -2733,16 +2723,3 @@ def test_dataset_drop(tmp_path: Path):
     assert Path(tmp_path).exists()
     lance.LanceDataset.drop(tmp_path)
     assert not Path(tmp_path).exists()
-
-
-@pytest.mark.integration
-def test_s3_drop(s3_bucket: str):
-    storage_options = copy.deepcopy(CONFIG)
-    table_name = uuid.uuid4().hex
-    tmp_path = f"s3://{s3_bucket}/{table_name}"
-    table = pa.table({"x": [0]})
-    dataset = lance.write_dataset(table, tmp_path, storage_options=storage_options)
-    dataset.validate()
-    lance.LanceDataset.drop(tmp_path, storage_options=storage_options)
-    with pytest.raises(Exception):
-        dataset.validate()
